+++ easyJavaDemo/src/main/java/com/easyjava/mappers/BaseMapper.java
@@ -6,19 +6,19 @@ public interface BaseMapper<T,P>{
     /*
     insert:插入
      */
-    Long insert(@Param("bean") T t);
+    Integer insert(@Param("bean") T t);
     /*
     插入或者更新
      */
-    Long insertOrUpdate(@Param("bean") T t);
+    Integer insertOrUpdate(@Param("bean") T t);
     /*
     批量插入
      */
-    Long insertBatch(@Param("list") List<T> list);
+    Integer insertBatch(@Param("list") List<T> list);
     /*
     批量插入或者更新
      */
-    Long insertOrUpdateBatch(@Param("list") List<T> list);
+    Integer insertOrUpdateBatch(@Param("list") List<T> list);
     /*
     根据参数查询集合
      */
@@ -26,5 +26,5 @@ public interface BaseMapper<T,P>{
     /*
     根据集合查询数量
      */
-    Long selectCount(@Param("query") P p);
+    Integer selectCount(@Param("query") P p);
 }
